Removing the nomath signed and unsigned functions
The signed and unsigned functions are trivial to do directly with >>0
and >>>0, and I don't trust inliners with this many levels of
indirection, so I'll just suffer the nonobvious code of >>0 and >>>0.
Perhaps a better, eventual solution would be to combine the C
preprocessor with JavaScript, but this'll do for now.